+++ test/ArrayTests.mjs
@@ -27,7 +27,31 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
+        7,
+        20,
+        42
+      ],
+      "getUnsafe - existing"
+    ], 1, eq, 1);
+
+Test.run([
+      [
+        "ArrayTests.res",
         8,
+        20,
+        41
+      ],
+      "getUnsafe - missing"
+    ], [
+        0,
+        1,
+        2
+      ][10], eq, undefined);
+
+Test.run([
+      [
+        "ArrayTests.res",
+        11,
         13,
         30
       ],
@@ -47,7 +71,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        14,
+        17,
         20,
         28
       ],
@@ -70,7 +94,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        15,
+        18,
         20,
         36
       ],
@@ -80,7 +104,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        18,
+        21,
         13,
         30
       ],
@@ -108,7 +132,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        24,
+        27,
         13,
         38
       ],
@@ -123,7 +147,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        31,
+        34,
         13,
         26
       ],
@@ -146,7 +170,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        36,
+        39,
         20,
         41
       ],
@@ -156,7 +180,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        39,
+        42,
         13,
         35
       ],
@@ -184,7 +208,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        45,
+        48,
         13,
         38
       ],
@@ -199,7 +223,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        51,
+        54,
         20,
         41
       ],
@@ -219,7 +243,7 @@ var arr = [
 Test.run([
       [
         "ArrayTests.res",
-        54,
+        57,
         13,
         31
       ],
@@ -229,7 +253,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        65,
+        68,
         13,
         24
       ],
@@ -255,7 +279,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        70,
+        73,
         20,
         42
       ],
@@ -274,7 +298,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        72,
+        75,
         13,
         32
       ],
@@ -289,7 +313,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        78,
+        81,
         20,
         30
       ],
@@ -306,7 +330,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        80,
+        83,
         13,
         34
       ],
@@ -324,7 +348,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        85,
+        88,
         20,
         41
       ],
@@ -338,7 +362,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        86,
+        89,
         20,
         38
       ],
@@ -348,7 +372,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        89,
+        92,
         13,
         22
       ],
@@ -370,7 +394,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        94,
+        97,
         20,
         40
       ],
@@ -389,7 +413,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        96,
+        99,
         13,
         30
       ],
@@ -404,7 +428,7 @@ Test.run([
 Test.run([
       [
         "ArrayTests.res",
-        103,
+        106,
         13,
         27
       ],