+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Add `Dict.getUnsafe` https://github.com/rescript-association/rescript-core/pull/167
 
+### Documentation
+
+- Fix docstring for `Array.getUnsafe` and `Array.filterMap` https://github.com/rescript-association/rescript-core/pull/168
+
 ## 0.6.0
 
 ### API changes